+++ packages/form-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/form-core
 
+## 1.24.4
+
+### Patch Changes
+
+- Optimise event client emissions and minor layout tweaks ([#1758](https://github.com/TanStack/form/pull/1758))
+
+## 1.24.3
+
+### Patch Changes
+
+- respect dontValidate option in formApi array modifiers ([#1775](https://github.com/TanStack/form/pull/1775))
+
+## 1.24.2
+
+### Patch Changes
+
+- fix(form-core): prevent runtime errors when using `deleteField` ([#1706](https://github.com/TanStack/form/pull/1706))
+
 ## 1.24.1
 
 ### Patch Changes